refactor : 카테고리
