+++ src/main/java/org/example/outsourcing_project/domain/menu/dto/response/MenuResponseDto.java
@@ -3,7 +3,7 @@ package org.example.outsourcing_project.domain.menu.dto.response;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.example.outsourcing_project.common.category.Category;
+import org.example.outsourcing_project.common.enums.Category;
 import org.example.outsourcing_project.domain.menu.entity.Menu;
 import org.example.outsourcing_project.domain.menu.entity.MenuOption;
